refactor: 댓글 코어 리팩토링 (#113)
