+++ module-core/src/main/java/com/back2basics/comment/port/in/CommentCreateUseCase.java
@@ -2,7 +2,7 @@ package com.back2basics.comment.port.in;
 
 import com.back2basics.comment.port.command.CommentCreateCommand;
 
-public interface CreateCommentUseCase {
+public interface CommentCreateUseCase {
 
     Long createComment(CommentCreateCommand command);
 }
