+++ module-core/src/main/java/com/back2basics/comment/port/in/CommentDeleteUseCase.java
@@ -1,6 +1,6 @@
 package com.back2basics.comment.port.in;
 
-public interface DeleteCommentUseCase {
+public interface CommentDeleteUseCase {
 
     void deleteComment(Long id, String requesterName);
 }
