+++ module-core/src/main/java/com/back2basics/comment/port/in/CommentUpdateUseCase.java
@@ -2,7 +2,7 @@ package com.back2basics.comment.port.in;
 
 import com.back2basics.comment.port.command.CommentUpdateCommand;
 
-public interface UpdateCommentUseCase {
+public interface CommentUpdateUseCase {
 
     void updateComment(Long id, CommentUpdateCommand command);
