+++ module-core/src/main/java/com/back2basics/comment/service/CommentUpdateService.java
@@ -1,0 +1,27 @@
+package com.back2basics.comment.service;
+
+import com.back2basics.comment.model.Comment;
+import com.back2basics.comment.port.command.CommentUpdateCommand;
+import com.back2basics.comment.port.in.CommentUpdateUseCase;
+import com.back2basics.comment.port.out.CommentUpdatePort;
+import com.back2basics.infra.validation.validator.CommentValidator;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class CommentUpdateService implements CommentUpdateUseCase {
+
+    private final CommentValidator commentValidator;
+    private final CommentUpdatePort commentUpdatePort;
+
+
+    @Override
+    public void updateComment(Long id, CommentUpdateCommand command) {
+        Comment comment = commentValidator.findComment(id);
+        commentValidator.isAuthor(comment, command.getRequesterName());
+
+        comment.update(command);
+        commentUpdatePort.update(comment);
+    }
+}
