+++ module-core/src/main/java/com/back2basics/infra/validation/validator/CommentValidator.java
@@ -1,7 +1,7 @@
 package com.back2basics.infra.validation.validator;
 
 import com.back2basics.comment.model.Comment;
-import com.back2basics.comment.port.out.CommentRepositoryPort;
+import com.back2basics.comment.port.out.CommentReadPort;
 import com.back2basics.service.comment.exception.CommentErrorCode;
 import com.back2basics.service.comment.exception.CommentException;
 import lombok.RequiredArgsConstructor;
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class CommentValidator {
 
-    private final CommentRepositoryPort commentRepository;
+    private final CommentReadPort commentRepository;
 
     public Comment findComment(Long id) {
         return commentRepository.findById(id)
